UI/requirements.txt
# Standalone UI Authentication System Requirements
# Complete list of packages needed to run the UI independently

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Password Hashing (specific version for compatibility)
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.0
